add execute permission for docker-manager scripts

+++ imtools/docker-manager/start-docker-manager.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! $(hostname) == 'docker-manager' ]; then
+	echo Failed. docker-manager should only run in host docker-manager!
+	exit
+fi
+
 echo Stop Compose project first if exists. 
 docker compose down
 echo Start Compose project.
